+++ delete.ts
@@ -1,5 +1,6 @@
 import FileDatabase from "./src/db/fileDatabase.js";
 import FileSystem from "./src/sys/fileSystem.js";
+import UserDatabase from "./src/db/userDatabase.js";
 
 const args = process.argv;
 if (args.length !== 3) {
@@ -12,8 +13,15 @@ else if (!args[2]) {
   process.exit(1);
 }
 
-const fileID = args[2];
+
+const userID = args[2];
+await UserDatabase.deleteUser(userID);
 Promise.all([
-  FileDatabase.deleteFile(fileID),
-  FileSystem.deleteFile(fileID)
+  async () => {
+    const files = await FileDatabase.listFiles(userID);
+    for (const file of files) {
+      await FileSystem.deleteFile(file.fileid);
+    }
+  },
+  UserDatabase.deleteUser(userID),
 ]).finally(() => process.exit(0));